Use HTTP/1.0 to contact Pinto
The upstream Pinto server has a bug in it: if you contact it with
HTTP/1.1, it generates a chunked response body, but never sends the
closing "0" chunk. HTTP::Tiny deals with that by reporting an
unsuccessful request. This isn't just an HTTP::Tiny issue, though: curl
yields essentially the same error (unless you explicitly downgrade to
HTTP/1.0).

Fix this by forcing the HTTP::Tiny request to be downgraded to HTTP/1.0
also. This is slightly annoying, because HTTP::Tiny doesn't offer that
feature: we have to subclass HTTP::Tiny to achieve it.